0️⃣ add suffixes to floating point literals
To enable compiling with `-Werror=unsuffixed-float-constants`, add some
suffixes to double literal constants.

Without using gnu extensions, there's no suffix available for double
literals, so instead use the suffix for `long double`, `L`, and cast to
explicit `double`. This should result in generated code equivalent to
using a double literal, even if precision loss occurs (it doesn't in
these cases though, and implicit conversion should be equivalent, but
just to be sure).

+++ include/CppUTestExt/MockCheckedActualCall.h
@@ -274,7 +274,7 @@ public:
     virtual unsigned int returnUnsignedIntValue() _override { return 0; }
     virtual unsigned int returnUnsignedIntValueOrDefault(unsigned int value) _override { return value; }
 
-    virtual double returnDoubleValue() _override { return 0.0; }
+    virtual double returnDoubleValue() _override { return (double)0.0L; }
     virtual double returnDoubleValueOrDefault(double value) _override { return value; }
 
     virtual const char * returnStringValue() _override { return ""; }
